Outline `Machine` concept
Am thinking through what exactly this is all about.  However, it
reflects what was done in EvmIL whereby we can separate abstract and
concrete semantics.  In particular, I think it would make sense to have
a refined notion `ConcreteMachine` which is for _runtime execution_.
This can then provide methods for executing (e.g. for a given number of
steps) the machine until an outcome is reached.